+++ src/o/crypto.ts
@@ -1,0 +1,1 @@
+export const crypto: typeof globalThis.crypto = import.meta.env.VSCODE_BROWSER_EXT ? globalThis.crypto : require('crypto').webcrypto;