+++ CHANGELOG.md
@@ -9,12 +9,14 @@ Users of the free version are not affected by this change.
 ### Secondary Sidebar
 SQLite Viewer now features a second sidebar 🎉. It shows either the current row selection or meta data about the current table.
 
+## v0.8.7
+Renamed "Enter Access Token" command to "Offline License Activation" and changed workflow to a guided wizard with additional instructions.
+
+
 ## v0.8.6
 Added "Enter Access Token" command that allows activating the PRO version while offline or behind a firewall. 
 
-Note that this is stop-gap solution and still requires _some_ communication to get the token (JWT) on the machine.
-
-To generate the Access Token, go to <https://vscode.sqliteviewer.app/api/register> and enter your license key. Currently, the token will be valid for 3 months.
+~~To generate the Access Token, go to https://vscode.sqliteviewer.app/api/register and enter your license key. Currently, the token will be valid for 3 months.~~
 
 ## v0.8.5
 ### Features